Renamed reference package to library
